Move AppMenuBar related components to designated folder
