+++ src/renderer/modules/AppMenuBarModule/constants.ts
