+++ src/renderer/modules/AppMenuBarModule/english.json
