+++ src/renderer/modules/AppMenuBarModule/korean.json
